feat : event struct was updated

+++ fw/Core/Inc/FSMs/ui/ui_draw.h
@@ -33,16 +33,16 @@ typedef enum
 
 typedef struct
 {
-    uint8_t x; 
-    uint8_t y; 
-    uint8_t w; 
-    uint8_t h;     
+    uint16_t x; 
+    uint16_t y; 
+    uint16_t w; 
+    uint16_t h;     
 }ui_window_t;
 
 typedef struct
 {   
-    uint8_t x; 
-    uint8_t y; 
+    uint16_t x; 
+    uint16_t y; 
     BITMAPSTRUCT *ptr;
 }ui_icon_t;
